Parameters for periodic boundary conditions
A new kind of boudnary is added to the boundary subsection for reading input files.
When periodic boundaruy condition is selected, the boundary id has to be given in
order to identify the boundary pair. The couple of periodic boundaries are stored in
a vector of integer vectors called periodic_pair.
Changes to be committed:
	modified:   src_mhd/parameters.cc
	modified:   src_mhd/parameters.h