Create configure_periodic_boundary function
Create an additional function in order to configurate periodic boudary
conditions. It uses the pairing of boudaries created when the parameters
are readed. It uses the fuctions:
 - collect_periodic_faces()
 - add_periodicity()
It was added at the run() function just before setup_system()
Warning: It has not been tested yet.
Todo: Implement a way to get the information at the boudary integration
      inside the MeshWorker loop.
integrate along the boundaries
 Changes to be committed:
	modified:   src_mhd/claw.cc
	modified:   src_mhd/claw.h