DealiiExtensions taken from dealii users forum
A library from dealii, modified by kraemer and again modified by me. It creates a map (cell,face) --> FacePair when Periodic Boundary conditions are selected. It is used for DG schemes where the periodic neighbor is needed in order to compute the numerical flux at the interface. This version only works for 2D.
 Changes to be committed:
	new file:   src_mhd/DealiiExtensions.cc
	new file:   src_mhd/DealiiExtensions.h