Modification of the Configuration of Periodic BC
A periodicity vector and a periodic map are created as vaiables in the ConservationLaw
object so one can access them in any process, particulary when the MeshWorker is
intergrating over the boundaries.

In the configurate_periodic_boundary() function the boundary_pair, which stores the pairing
 of periodic boundaries was changed from vector<vector<boundary_id>> to vector<pair<boundary_id>>

The make_periodicity_map_dg() is called to create the map between corresponding cells
at the periodic boundary.

Finally, a mapping correction line was added for dealii-8.4 compatibility