Periodic Boundary Conditions for dflo
New parameters were defined in order to read periodic boudary
conditions from a configuration file. The input parameters should
include the corresponding pair boundary (pair) and the direction
of the periodicity (ether x or y).

In the run function from the claw.cc file a new function is called
to configurate periodic boundaries. it call routines from the
DealiiExtensions files.

The assemble_explicit.cc file now includes an extension in the
boundary intergration. When periodic boundaries are present
it finds the data for the corresponding periodic "neighbor" and
computes the numerical fluxes.

The equation.h file just has a new tipe of boundary condition
called "periodic"