Periodic boundary conditions (#24)
* minor changes

* Working version without refinement

* revert input file changes

* revert input file changes

* Time consuption flags

* my changes to claw.cc file before merging

* correct indentation, minor changes

* using dealii parallel distributed vectors

* eliminating useless files

* correcting indentation

* MHD directory

* MHD directory and MHD structure

* Artificial Reflective Boundary Condition

* Point fix for change in deal.ii

* fix bug Point

* Fix for dealii version 8.3

* Fix for dealii version 8.3

* Add multid folder from Praveens repository

* Remove old implementation of MHD from master branch

* Eliminate double from Tensor definitions

* add fodler for MHD implementation

* Update old src_mhd folder to new src_mpi

In order to start implementing the MHD code it is
usefull to add the MPI current status to the old
MHD begining of the implementation

* Work in progress: Updating the MHD implementation

Adding implementations for the Alfven, slow and fast wave
to the eigenvalues and eigenvectors can be calculated easily.
The implementation of the maximum characteristic speed is
straightforward and it is also included.

TODO: Check if the Alfven speed is correct, or if it changes
for higher dimensions.

* =Fix path of libraries and Tensor instead of points in src

* external_force problem in parameters.cc fix and integrator flags in claw.cc

* Adding MHD equations for finite volume test

* Parameters for periodic boundary conditions

A new kind of boudnary is added to the boundary subsection for reading input files.
When periodic boundaruy condition is selected, the boundary id has to be given in
order to identify the boundary pair. The couple of periodic boundaries are stored in
a vector of integer vectors called periodic_pair.
Changes to be committed:
	modified:   src_mhd/parameters.cc
	modified:   src_mhd/parameters.h

* Add periodic boundary conditions to the list

Add periodic boundary conditions to the list in the Equation object.
Changes to be committed:
	modified:   src_mhd/equation.h

* Create configure_periodic_boundary function

Create an additional function in order to configurate periodic boudary
conditions. It uses the pairing of boudaries created when the parameters
are readed. It uses the fuctions:
 - collect_periodic_faces()
 - add_periodicity()
It was added at the run() function just before setup_system()
Warning: It has not been tested yet.
Todo: Implement a way to get the information at the boudary integration
      inside the MeshWorker loop.
integrate along the boundaries
 Changes to be committed:
	modified:   src_mhd/claw.cc
	modified:   src_mhd/claw.h

* Boundary integrator with periodic BC

In the MeshWorker functions, a periodic boundary condition chech is added for the boundary integration. If the boundary happen to be periodic, it will load the face information from the corresponding periodic boundary neighbor and perform the integration as in the case of cell interface integration. The information is taken from a (cell0,face0)-->face_pair mapping from the DealiiExtensions files.
 Changes to be committed:
	modified:   src_mhd/assemble_explicit.cc

* DealiiExtensions taken from dealii users forum

A library from dealii, modified by kraemer and again modified by me. It creates a map (cell,face) --> FacePair when Periodic Boundary conditions are selected. It is used for DG schemes where the periodic neighbor is needed in order to compute the numerical flux at the interface. This version only works for 2D.
 Changes to be committed:
	new file:   src_mhd/DealiiExtensions.cc
	new file:   src_mhd/DealiiExtensions.h

* Modification of the Configuration of Periodic BC

A periodicity vector and a periodic map are created as vaiables in the ConservationLaw
object so one can access them in any process, particulary when the MeshWorker is
intergrating over the boundaries.

In the configurate_periodic_boundary() function the boundary_pair, which stores the pairing
 of periodic boundaries was changed from vector<vector<boundary_id>> to vector<pair<boundary_id>>

The make_periodicity_map_dg() is called to create the map between corresponding cells
at the periodic boundary.

Finally, a mapping correction line was added for dealii-8.4 compatibility

* Update reading parameters for Periodic BC

Correspondent periodic boundary conditions are saved in a vector of pairs. It is ensure that the pairing is saved just once.

* Check if orientation is the same in the PBC faces

If the orientation of the faces, given by face_pair.orientation[1],
is different then the order of the quadrature points is inverted.

* Integration in periodic boudaries for MeshWorker

An modification in the integration over boundaries has been introduced
on order to take into account periodic boundaries. The information of
the corresponding periodic element is loaded as a neighbor, then the
integration is performed as in the case of the inter element interfaces.

* Paraeters necesary for Periodic Boundary Condition

Parameters like periodic type, boundary pair and direction were
added to the parameters files in order to admit test cases with
periodic boundary conditions.

* Configure periodic boundary conditions direction

Direction was added to the PBC configuration in order to admit periodicity
both in x and y direction. It just works for 2D, however can be extended
to 3D by modifying the parameter reading files.

* Change DealExtensions to use pararell::Distributed

Change the iterator of the DealiiExtensions file to use
parallel::distributed::Triangulation::cell_iterator which
is present in our implementation.

* Minor changes

* Periodic Boundary Conditions for dflo

New parameters were defined in order to read periodic boudary
conditions from a configuration file. The input parameters should
include the corresponding pair boundary (pair) and the direction
of the periodicity (ether x or y).

In the run function from the claw.cc file a new function is called
to configurate periodic boundaries. it call routines from the
DealiiExtensions files.

The assemble_explicit.cc file now includes an extension in the
boundary intergration. When periodic boundaries are present
it finds the data for the corresponding periodic "neighbor" and
computes the numerical fluxes.

The equation.h file just has a new tipe of boundary condition
called "periodic"

* Prepare periodic BC for commit on cpraveen-master

* Adding DealiiExtensions Files

* Changes to cpraveen-master just in src_mpi

* Vortex test case for Periodic BC

An adaptation of the vortex sistem for periodic boundary conditions
was added. It shows how the PBC should be configured.

* Minor change

* Updating the vector of neighbor for PBC

In the setup_system() function, the assignment of neighbors
needed to be changed, so the limiter will work also through
boundaries tagged as periodic.

* Minor changes

* Minor changes

* minor changes